refactor: create file for getHoseCodesForAtomsAsFragments
