+++ src/path/getShortestPaths.js
@@ -32,8 +32,7 @@ export function getShortestPaths(molecule, options = {}) {
         molecule.getPath(path, from, to, maxLength);
         if (path.length > 0) {
           allShortestPaths[from][to] = path.slice();
-          // eslint-disable-next-line unicorn/no-array-reverse
-          allShortestPaths[to][from] = path.reverse();
+          allShortestPaths[to][from] = path.toReversed();
         } else {
           allShortestPaths[from][to] = null;
           allShortestPaths[to][from] = null;
